refactor downscaling code structure
