+++ oneflux/downscaling/gapfilling.py
@@ -1,166 +1,301 @@
-#!/usr/bin/env python
-# command line to run:
-#   python gapfilling.py config_all.txt    
-import sys
-import re
 import numpy as N
-import os
 import logging
 
-from oneflux import ONEFluxError
-from oneflux.downscaling.functions import search_index, solarang, time_zone, residuals, rmean, \
-                                          read_weather_papale, read_ERA5_da_clusterFR,write_csv
-from oneflux.downscaling.gap_fill import *
-from oneflux.downscaling.write_stat import *
+from scipy.optimize import leastsq
+from scipy import stats
+from oneflux.downscaling.functions import solarang, time_zone, residuals, rms
+from oneflux.downscaling.constantes import * 
 
 log = logging.getLogger(__name__)
 
-def read_config(file_config):
-    # importo il file di configurazione
-    with open(file_config) as f:
-        for line in f.readlines():
-            if line.startswith('#'):
-                continue
-            line = line.strip()
-            if len(line) == 0:
-                continue
-            line = line.split('#')[0]
-            if len(line) == 0:
-                continue
-            line = line.split('=')
-            line[0] = line[0].strip()
-            line[1] = line[1].strip()
-            log.debug('Config file entries: {l1} = {l2}'.format(l1=line[0], l2=line[1]))
-            if line[0] == 'name_path_weather':
-                name_path_weather = line[1]
-            if line[0] == 'name_path_reanalysis':
-                name_path_reanalysis = line[1]
-            if line[0] == 'name_path_out':
-                name_path_out = line[1]
-            if line[0] == 'gapmax':
-                gapmax = float(line[1])
-            if line[0] == 'Site':
-                name = line[1]
-            if line[0] == 'FirstY':
-                year_start = int(line[1])
-            if line[0] == 'LastY':
-                year_end = int(line[1])
-            if line[0] == 'Lat':
-                lat = float(line[1])
-            if line[0] == 'Lon':
-                lon = float(line[1])
-            if line[0] == 'UTCtime':
-                timeshift = float(line[1])
-            if line[0] == 'timeres':
-                weather_period = float(line[1])
-            if line[0] == 'pixel':
-                pixel = line[1]
-                
-    return {'name_path_weather': name_path_weather,
-         'name_path_reanalysis': name_path_reanalysis,
-         'name_path_out': name_path_out,
-         'gapmax': gapmax,
-         'Site': name,
-         'FirstY': year_start,
-         'LastY': year_end,
-         'Lat': lat,
-         'Lon': lon,
-         'UTCtime': timeshift,
-         'timeres': weather_period,
-         'pixel': pixel
-        }
-def gapfilling(file_config):
+#-------------------------------------------------------------------------------------------
+# function gap_fill_func gapfills the meteorlogical data
+# argument1 : the weather dataset (weather)
+# argument2 : the climatology dataset (clim)
+# argument3 : the ratio between clim and weather time steps (diff_clim_weather)
+# argument4 : the weather time step
+# argument5 : the clim time step
+# argument6 : vector of julian days
+# argument7 : length of the year
+# argument8 : longitude
+# argument9 : latitude
+# argument10: maximum length of gap below which we linearly interpolated
+# argument11: vector of flags indicating whether the field is averaged or instantaneous
+# argument12: vector of time step shifted needed for putting in agreement weather and climate datasets
+# argument13: vector of time step shifted needed for accounting for time zone adjustment
+ 
+# returns a vector that contains the weather dataset gapfilled
+def gap_fill_func(weather,clim,diff_clim_weather,weather_period,climato_period,julian,year_length,lon,lat,gapmax,avg,climatoshift,timeshift):
+   weather_clim_period=[]
+   weather_clim_period_test=[]
 
-    dict_config = read_config(file_config)
-    id_lwdown_calc=7
-    id_snow=8
-    
-    log.debug('Treatment for site: {s}'.format(s=dict_config['Site']))
-    for k,v in dict_config.items():
-        log.debug('Entry [{k}] = {v}'.format(k=k, v=v))
+   weather_clim_period_nogap=[]
+   clim_nogap=[]
+   weather_clim_period_gapfill=[]
+   weather_clim_period_all_gapfill=[]
+   weather_gapfill=[]
+   weather_all_gapfill=[]
+   clim_weather_period=[]
 
-    clim_dict = read_ERA5_da_clusterFR(dict_config['Site'],year_start=dict_config['FirstY'],year_end=dict_config['LastY'],
-                                       dir_input_era5=dict_config['name_path_reanalysis'],pixel=dict_config['pixel'])
-                                       #lon_d = 0,lat_d=0,
-                                       #dataset_name=''):
-    clim = clim_dict['clim']
-    log.debug('clim: {s}'.format(s=clim))
-    climato_period = clim_dict['climato_period']
-    avg = clim_dict['avg']
-    climatoshift = clim_dict['climatoshift']
-    
-    for k in range(len(clim)):
-       clim[k]=N.array(clim[k],float)
-    # Time frequency of clim data (in hours)
-    #   climato_period is set when calling read_climatology function
-    diff_clim_weather=(int)(climato_period/dict_config['timeres'])
-    
-    #------------------------------------------------------------------------------------------- 
-    # Download of the weather variables from FLUXNET dataset
-    # for the i_th site stored in the site TXT file
-    year=dict_config['FirstY']
-    filename=dict_config['name_path_weather']+dict_config['Site']+'_'+str(year)+'_synth_hourly_allvars.csv'
-    weather=read_weather_papale(dict_config['name_path_weather'],dict_config['Site'],1,dict_config['timeres'],str(year))
-    julian=[]
-    year_length=[]
-    julian.extend((range(len(weather[0]))+N.ones(1)/2.)/N.array(24./dict_config['timeres']))
-    year_length.extend(range(len(weather[0]))*N.array(0.)+N.array(len(weather[0])/(24./dict_config['timeres'])))
-    
-    for j in range(dict_config['LastY']-dict_config['FirstY']):
-       year=dict_config['FirstY']+j+1
-       filename=dict_config['name_path_weather']+dict_config['Site']+'_'+str(year)+'_synth_hourly_allvars.csv'
-       weather_year=read_weather_papale(dict_config['name_path_weather'],dict_config['Site'],1,dict_config['timeres'],str(year))
-       for k in range(len(weather_year)):      
-          weather[k].extend(weather_year[k])
-       julian.extend((range(len(weather_year[0]))+N.ones(1)/2.)/N.array(24./dict_config['timeres']))
-       year_length.extend(range(len(weather_year[0]))*N.array(0.)+len(weather_year[0])/(24./dict_config['timeres']))
-    
-    for k in range(len(weather)):
-       weather[k]=N.array(weather[k],float)
-       
-    julian=N.array(julian)
-    year_length=N.array(year_length)
-    
-    weather_qc=N.array(weather)
-    weather_qc=N.where(weather_qc==-9999,0,1)
-    
-    if (len(weather[0])/len(clim[0]) != (climato_period/dict_config['timeres'])):
-       log.debug('For {s}'.format(s=dict_config['Site']))
-       msg = 'Timelength of Clim and Weather datasets do not match'
-       log.critical(msg)
-       raise ONEFluxError(msg)
-    
-    
-    #-------------------------------------------------------------------------------------------
-    # Fill the gaps in the meteorological fields
-    [clim_weather_period,weather_all_gapfill,weather_gapfill,weather_clim_period_gapfill,weather_clim_period_nogap,
-     weather_clim_period,clim_nogap,stat]=gap_fill_func(weather,clim,diff_clim_weather,dict_config['timeres'],climato_period,
-                                                        julian,year_length,dict_config['Lon'],dict_config['Lat'],dict_config['gapmax'],avg,climatoshift,dict_config['UTCtime'])
-    
-    if not os.path.exists(dict_config['name_path_out']):
-        os.makedirs(dict_config['name_path_out'])
-    
-    write_stat(weather,clim_nogap,weather_clim_period_nogap,dict_config['UTCtime'],dict_config['timeres'],dict_config['name_path_out'],dict_config['Site'])
-    
-    write_stat_30min(weather,weather_all_gapfill,dict_config['UTCtime'],dict_config['timeres'],dict_config['name_path_out'],dict_config['Site'])
-    write_stat_30min(weather,clim_weather_period,dict_config['UTCtime'],dict_config['timeres'],dict_config['name_path_out'],dict_config['Site']+'_nocorr')
-    
-    write_csv(weather_all_gapfill,dict_config['FirstY'],dict_config['LastY'],dict_config['timeres'],dict_config['Site'],dict_config['name_path_out'])
-    
-    write_csv(clim_weather_period,dict_config['FirstY'],dict_config['LastY'],dict_config['timeres'],dict_config['Site']+'_nocorr',dict_config['name_path_out'])
+   if(timeshift==-9999):
+      timezone=time_zone(0,lon)
+      # east of Greenwich => timeshift>0
+      # west of Greenwich => timeshift<0
+      if(timezone[1]<13):
+         timeshift=timezone[1]
+      else:
+         timeshift=timezone[1]-24
 
-    log.debug(dict_config['name_path_out'])
+   stat_vec=[]
+
+   log.debug('shift to UTC time = {t} hours'.format(t=timeshift))
 
 
-def main(argv):
-    args = argv[1:]
-    log.debug(args)
-    #start_process_total = datetime.datetime.now()
-    gapfilling(args[0])
-    #end_process_total = datetime.datetime.now()
-    #elap_process_total = end_process_total - start_process_total           
+   for k in range(len(weather)):
+   
+      #weather_clim_period.append(N.zeros(int(len(weather[k])/diff_clim_weather), dtype='float', 0))
+      weather_clim_period.append(N.zeros(int(len(weather[k])/diff_clim_weather), dtype=float))
+      if(k==id_precip):
+         #freq_precip=N.zeros(len(weather[k])/diff_clim_weather, float, 0)
+         freq_precip=N.zeros(int(len(weather[k])/diff_clim_weather), float)
+      #weather_clim_period_test.append(N.zeros(len(weather[k])/diff_clim_weather, float, 0))
+      weather_clim_period_test.append(N.zeros(int(len(weather[k])/diff_clim_weather), dtype=float))
 
 
-if __name__ == "__main__":
-    log.debug(sys.argv)
-    main(sys.argv)
+      # climatoshift indicates to which time step or time period
+      # unit is fraction of a time period (between 2 consecutive time steps)
+      # a climatic field value corresponds to
+      # for field that is a mean value (avg=1)
+      #    climatoshift=0 when the mean value is calculated from one time step to the next one
+      #    climatoshift=-0.5 when the mean value is centered on one time step
+      # for field that is an instantaneous value (avg=0)
+      #    climatoshift=0 when the value correspond to the current time step
+      #    climatoshift=1 when the value corresponds to the next time step
+      totalshift=timeshift+climatoshift[k]*climato_period
+      log.debug('Variable being processed: {v}'.format(v=label_fig[k]))
+      for t in range(len(weather[k])):
+         tshift=t-(float)(totalshift)/weather_period
+         goodcell=int(tshift/diff_clim_weather)
+         if((tshift>=0)and(tshift<len(weather[k]))):
+            # in case of a mean value calculation, we sum all weather element within each element of weather_clim_period
+            # if one weather element equals -9999, the related weather_clim_period is equal to -9999
+            if (avg[k]==1):
+               if((weather[k][t] != -9999) and (weather_clim_period[k][goodcell] != -9999)):
+                  weather_clim_period[k][goodcell]+=weather[k][t]/diff_clim_weather
+               else:
+                  weather_clim_period[k][goodcell]=-9999
+               weather_clim_period_test[k][goodcell]=1
+
+               if(k==id_precip):
+                  if((weather[k][t] != -9999) and (freq_precip[goodcell] != -9999)):
+                     if(weather[k][t]>0):
+                        freq_precip[goodcell]+=1./diff_clim_weather
+                  else:
+                     freq_precip[goodcell]=-9999
+               
+            # in case of a instantaneaous calculation, each weather_clim_period element corresponds
+            # to the first weather element associated to this weather_clim_period element
+            else:
+               if(weather_clim_period_test[k][goodcell] == 0):
+                  if((weather[k][t] != -9999)):
+                     weather_clim_period[k][goodcell]=weather[k][t]
+                  else:
+                     weather_clim_period[k][goodcell]=-9999
+                  weather_clim_period_test[k][goodcell]=1
+      # in case of a mean value calculcation (avg==1)
+      # elements of weather_clim_period that have been partly filled (at the beginning or at the end)
+      # are set to -9999
+      if (avg[k]==1):
+         if(totalshift%climato_period !=0):
+            if(totalshift>0):
+               goodcellmissing=int((len(weather[k])-1-(float)(totalshift)/weather_period)/diff_clim_weather)
+               weather_clim_period[k][goodcellmissing]=-9999
+            else:
+               if(totalshift<0):
+                  goodcellmissing=int((-(float)(totalshift)/weather_period)/diff_clim_weather)
+                  weather_clim_period[k][goodcellmissing]=-9999
+      # elements of weather_clim_period that have NOT been filled (at the beginning or at the end)
+      # are set to -9999          
+      weather_clim_period[k]=N.where(weather_clim_period_test[k]==1,weather_clim_period[k],-9999)
+               
+
+      weather_clim_period_nogap.append([])
+      clim_nogap.append([])
+      for t in range(len(weather_clim_period[k])):
+         if(weather_clim_period[k][t] != -9999):
+            weather_clim_period_nogap[k].append(weather_clim_period[k][t])
+            clim_nogap[k].append(clim[k][t])
+
+
+      weather_clim_period_nogap[k]=N.array(weather_clim_period_nogap[k],float)
+      clim_nogap[k]=N.array(clim_nogap[k],float)
+
+
+
+      # Evaluate the correlation
+      # between clim_nogap and weather_clim_period_nogap
+      # intercept and slope of the relation
+      # will be used for correcting clim data when filling gaps
+      a=clim_nogap[k]
+      b=weather_clim_period_nogap[k]
+
+      if((k==id_swdown)or(k==id_ws)):
+         if(len(a)==0):
+            stat=N.zeros(2)
+            stat[0]='nan'
+         else:
+            stat1=leastsq(residuals,1.,args=(b,a))
+            stat=N.zeros(2, dtype=float)
+            #stat=N.array(stat[0], dtype=float)
+            stat[0]=stat1[0][0]
+            stat[1]=0
+      elif(k==id_precip):
+         stat=N.zeros(2, dtype=float)
+         if(len(a)==0):
+            stat[0]='nan'
+            stat[1]=0
+         else:
+            stat[0]=sum(b)/sum(a)
+            stat[1]=0
+      else:
+         if(len(a)==0):
+            stat=N.zeros(2, dtype=float)
+            stat[0]='nan'
+         else:
+            #stat=statistics.linearregression(b,x=a)
+            slope, intercept, r, p, std_err = stats.linregress(a, b)
+            stat=N.zeros(2)
+            stat[0] = slope
+            stat[1] = intercept 
+      if(str(stat[0])!='nan'):
+         slope=stat[0]
+         intercept=stat[1]
+      else:
+         slope=1
+         intercept=0
+
+      stat_vec.append([slope,intercept])
+      log.debug('Slope of the linear relation in-situ VS reanalysis={s}'.format(s=slope))
+      log.debug('Intercept of the linear relation in-situ VS reanalysis={s}'.format(s=intercept))
+      if(str(stat[0])!='nan'):
+         log.debug('RMSE without correction={s}'.format(s=rms(a,b)))
+         log.debug('RMSE with correction={s}'.format(s=rms(a*slope+intercept,b)))
+         
+
+      weather_clim_period_gapfill.append([])
+      weather_clim_period_gapfill[k]=N.where(weather_clim_period[k]==-9999, slope*clim[k]+intercept, weather_clim_period[k])
+
+      weather_clim_period_all_gapfill.append([])
+      weather_clim_period_all_gapfill[k]=slope*clim[k]+intercept
+
+      weather_all_gapfill.append([])
+      clim_weather_period.append([])
+ 
+      weather_gapfill.append([])
+      mean_csang=0.
+      if(k==id_precip):
+         freq_precip_nogap=N.ma.masked_values(freq_precip,-9999)
+         freq_precip_nogap_nonull=N.ma.masked_values(freq_precip_nogap,0.)
+         if(freq_precip_nogap_nonull.count()!=0):
+            freq_precip_scalar=freq_precip_nogap_nonull.mean()
+            number_precip_per_diff_clim_weather=round(freq_precip_scalar*diff_clim_weather)
+         else:
+            number_precip_per_diff_clim_weather=diff_clim_weather
+      for t in range(len(weather[k])):
+         tshift=t-(float)(totalshift)/weather_period
+         goodcell=int(tshift/diff_clim_weather)
+         if(k==id_swdown):
+            if((t-(float)(timeshift)/weather_period)<(len(weather[k])-(diff_clim_weather-1))):
+               if(tshift%diff_clim_weather==0):
+                  mean_csang=0.
+                  for l in range(diff_clim_weather):
+                     goodcellsolar=int((t-(float)(timeshift)/weather_period)+l)
+                     mean_csang+=solarang(julian[goodcellsolar],0,lon,lat,year_length[int(tshift)])/diff_clim_weather
+               if(mean_csang == 0.):
+                  ratio=0.
+               else:
+                  goodcellratio=int((t-(float)(timeshift)/weather_period))
+                  ratio=solarang(julian[goodcellratio],0,lon,lat,year_length[goodcellratio])/mean_csang
+            else:
+               ratio=1
+         if(k==id_lwdown):
+            ratio=1
+         if(k==id_precip):
+            if(t%diff_clim_weather+1<=number_precip_per_diff_clim_weather):
+               ratio=diff_clim_weather/number_precip_per_diff_clim_weather
+            else:
+               ratio=0.
+
+         # if the current value is -9999, we have to fill the gap
+         if(weather[k][t] == -9999):
+            # if gapmax is > 0
+            # for short gap, we will try to interpolate
+            # between the last and the next
+            # defined values in the weather dataset
+            active_linear=1
+            if(gapmax>0):
+               tlow=t-1
+               while((tlow>=0)and(weather[k][tlow]==-9999)and((t-tlow)<=gapmax)):
+                  tlow=tlow-1
+               tup=t+1
+               while((tup<len(weather[k]))and(weather[k][tup]==-9999)and((tup-t)<=gapmax)):
+                  tup=tup+1
+               # if the last and the next defined values are not too far
+               # we linearly interpolate with the weather dataset
+               if(tup-tlow<=gapmax+1):
+                  if(tlow<0):
+                     weather_gapfill[k].append(weather[k][tup])
+                  elif(tup>=len(weather[k])):
+                     weather_gapfill[k].append(weather[k][tlow])
+                  else:
+                     step=(float)(t-tlow)/(tup-tlow)
+                     weather_gapfill[k].append(weather[k][tlow]*(1-step)+weather[k][tup]*step)
+               else:
+                  active_linear=0
+            # if gapmax == 0 or
+            # if the last and the next defined values are too far each other
+            # we use the clim dataset for filling the gap
+            if((gapmax==0)or(not active_linear)):   
+               if (avg[k]==0):
+                  step=(float)(tshift%diff_clim_weather)/diff_clim_weather
+                  if(tshift<0.):
+                     weather_gapfill[k].append(weather_clim_period_gapfill[k][0])
+                  elif((tshift/diff_clim_weather+1)>=(len(weather_clim_period_gapfill[k]))):
+                     weather_gapfill[k].append(weather_clim_period_gapfill[k][len(weather_clim_period_gapfill[k])-1])
+                  else:
+                     weather_gapfill[k].append(weather_clim_period_gapfill[k][goodcell]*(1-step)+weather_clim_period_gapfill[k][goodcell+1]*step)
+               else:
+                  if(tshift<0.):
+                     weather_gapfill[k].append(weather_clim_period_gapfill[k][0]*ratio)
+                  elif((tshift/diff_clim_weather)>=(len(weather_clim_period_gapfill[k]))):
+                     weather_gapfill[k].append(weather_clim_period_gapfill[k][len(weather_clim_period_gapfill[k])-1]*ratio)
+                  else:
+                     weather_gapfill[k].append(weather_clim_period_gapfill[k][goodcell]*ratio)
+         else:
+            weather_gapfill[k].append(weather[k][t])
+
+         if (avg[k]==0):
+            step=(float)(tshift%diff_clim_weather)/diff_clim_weather
+            if(tshift<0.):
+               weather_all_gapfill[k].append(weather_clim_period_all_gapfill[k][0])
+               clim_weather_period[k].append(clim[k][0])
+            elif((tshift/diff_clim_weather+1)>=(len(weather_clim_period_all_gapfill[k]))):
+               weather_all_gapfill[k].append(weather_clim_period_all_gapfill[k][len(weather_clim_period_all_gapfill[k])-1])
+               clim_weather_period[k].append(clim[k][len(clim[k])-1])
+            else:
+               weather_all_gapfill[k].append(weather_clim_period_all_gapfill[k][goodcell]*(1-step)+weather_clim_period_all_gapfill[k][goodcell+1]*step)
+               clim_weather_period[k].append(clim[k][goodcell]*(1-step)+clim[k][goodcell+1]*step)
+         else:
+            if(tshift<0.):
+               weather_all_gapfill[k].append(weather_clim_period_all_gapfill[k][0]*ratio)
+               clim_weather_period[k].append(clim[k][0]*ratio)
+            elif((tshift/diff_clim_weather)>=(len(weather_clim_period_all_gapfill[k]))):
+               weather_all_gapfill[k].append(weather_clim_period_all_gapfill[k][len(weather_clim_period_all_gapfill[k])-1]*ratio)
+               clim_weather_period[k].append(clim[k][len(clim[k])-1]*ratio)
+            else:
+               weather_all_gapfill[k].append(weather_clim_period_all_gapfill[k][goodcell]*ratio)
+               clim_weather_period[k].append(clim[k][goodcell]*ratio)
+
+   return clim_weather_period,weather_all_gapfill,weather_gapfill,weather_clim_period_gapfill,weather_clim_period_nogap,weather_clim_period,clim_nogap,stat_vec
+# end function
+#-------------------------------------------------------------------------------------------
